ContinuumMechanicsBase

ContinuumMechanicsBase contains the core functionality for continuum mechanics methods. The methods are intended to be overloaded by the package implementing the common interface. The objective of the package is to provide a common interface between interacting continuum modelling packages. The package is currently used by TRACER Lab for multi-scale modelling of soft materials. Packages which rely on ContinuumMechanicsBase include:

Hyperelastics.jl

Further packages are in development and will be made public upon completion. Please submit a PR if you are using this package in your project.

ContinuumMechanicsBase.MaterialHistory

ContinuumMechanicsBase.AlmansiStrainTensor

ContinuumMechanicsBase.AlmansiStrainTensor!

ContinuumMechanicsBase.CauchyStressTensor

ContinuumMechanicsBase.CauchyStressTensor!

ContinuumMechanicsBase.DeformationGradientTensor

ContinuumMechanicsBase.DeformationGradientTensor!

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor!

ContinuumMechanicsBase.GreenStrainTensor

ContinuumMechanicsBase.GreenStrainTensor!

ContinuumMechanicsBase.InverseDeformationGradientTensor

ContinuumMechanicsBase.InverseDeformationGradientTensor!

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor!

ContinuumMechanicsBase.I₁

ContinuumMechanicsBase.I₂

ContinuumMechanicsBase.I₃

ContinuumMechanicsBase.J

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor!

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor!

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor!

ContinuumMechanicsBase.StrainEnergyDensity

ContinuumMechanicsBase.StrainEnergyDensity!

ContinuumMechanicsBase.VelocityGradientTensor

ContinuumMechanicsBase.VelocityGradientTensor!

ContinuumMechanicsBase.predict

ContinuumMechanicsBase.MaterialHistory — Type

Structure for storing the behavior of a material as it evolves in time. Design to be used in time-dependent models such as viscoelasticity.

ContinuumMechanicsBase.AlmansiStrainTensor! — Method

AlmansiStrainTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.AlmansiStrainTensor — Method

AlmansiStrainTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.CauchyStressTensor! — Method

CauchyStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.DeformationGradientTensor! — Method

DeformationGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.DeformationGradientTensor — Method

DeformationGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor! — Method

FirstPiolaKirchoffStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor — Method

FirstPiolaKirchoffStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.GreenStrainTensor! — Method

GreenStrainTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.GreenStrainTensor — Method

GreenStrainTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseDeformationGradientTensor! — Method

InverseDeformationGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseDeformationGradientTensor — Method

InverseDeformationGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor! — Method

InverseLeftCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor — Method

InverseLeftCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.I₁ — Method

I₁(T::AbstractMatrix) -> Any

$I_1 = \text{tr}(T)$

ContinuumMechanicsBase.I₂ — Method

I₂(T::AbstractMatrix) -> Any

$I_2 = \frac{1}{2}\left(\text{tr}(T)^2 - \text{tr}(T^2)\right)$

ContinuumMechanicsBase.I₃ — Method

I₃(T::AbstractMatrix) -> Any

$I_3 = \det{T}$

ContinuumMechanicsBase.J — Method

J(T::AbstractMatrix) -> Any

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor! — Method

LeftCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor — Method

LeftCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor! — Method

RightCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor — Method

RightCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor! — Method

SecondPiolaKirchoffStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.StrainEnergyDensity! — Method

StrainEnergyDensity!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.VelocityGradientTensor! — Method

VelocityGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.VelocityGradientTensor — Method

VelocityGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.predict — Method

predict(
    ψ::ContinuumMechanicsBase.AbstractMaterialModel,
    test::ContinuumMechanicsBase.AbstractMaterialTest,
    ps;
    kwargs...
)

Fields:

ψ: Material Model

test or tests: A single test or vector of tests. This is used to predict the response of the model in comparison to the experimental data provided.

ps: Model parameters to be used.