ContinuumMechanicsBase

ContinuumMechanicsBase contains the core functionality for continuum mechanics methods. The methods are intended to be overloaded by the package implementing the common interface. The objective of the package is to provide a common interface between interacting continuum modelling packages. The package is currently used by TRACER Lab for multi-scale modelling of soft materials. Packages which rely on ContinuumMechanicsBase include:

Hyperelastics.jl

Further packages are in development and will be made public upon completion. Please submit a PR if you are using this package in your project.

ContinuumMechanicsBase.MaterialHistory

ContinuumMechanicsBase.predict

ContinuumMechanicsBase.MaterialHistory — Type

MaterialHistory(values::Vector, times::Vector)

Structure for storing the behavior of a material as it evolves in time. Design to be used in time-dependent models such as viscoelasticity.

ContinuumMechanicsBase.predict — Method

predict(ψ::AbstractMaterialModel, test::AbstractMaterialTest, ps)

Fields:

ψ: Material Model

test or tests: A single test or vector of tests. This is used to predict the response of the model in comparison to the experimental data provided.

ps: Model parameters to be used.